+++ Runtime/Scripts/UnityPlatform/UnityPlatformApplication.cs
@@ -9,7 +9,7 @@ namespace Sendbird.Chat
 {
     internal class UnityPlatformApplication : IPlatformApplication
     {
-        string IPlatformApplication.SdkVersion => "4.0.1";
+        string IPlatformApplication.SdkVersion => "4.1.0";
         string IPlatformApplication.PlatformName => "Unity";
         string IPlatformApplication.PlatformVersion => Application.unityVersion;
         string IPlatformApplication.OsName => Application.platform.ToString();