+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Change Log
 
-## 4.0.1 (Nov 4, 2024)
+## 4.1.0 (Nov 29, 2024)
 ### Features
+- Added `SetPushTriggerOption` to `SendbirdChatClient`
+- Added `GetPushTriggerOption` to `SendbirdChatClient`
+- Added `SetMyPushTriggerOption` to `SbGroupChannel`
+- Added `GetMyPushTriggerOption` to `SbGroupChannel`
+- Added `SbPushTriggerOption`
+### Bug Fixes
+- Fixed an issue with `SendbirdChat.BlockUser` where 'User not found error' occurs due to URL encoding
+
+## 4.0.1 (Nov 4, 2024)
+### Bug Fixes
 - Fixed an issue where build failed on the Windows platform
 
 ## 4.0.0 (Sep 25, 2024)
@@ -19,7 +29,7 @@
 - Added support for .NET 4.x
 
 ## 4.0.0-beta.1 (Sep 15, 2023)
-### Improvements
+### Bug Fixes
  - Fixed the bug regarding the URL encoding
 
 ## 4.0.0-beta (Aug 25, 2023)